// src/common/styles/palette.d.ts
export declare const palette: {
    green: {
        300: string;
        600: string;
        800: string;
    };
    white: {
        300: string;
        600: string;
        800: string;
    };
    grey: {
        200: string;
        300: string;
        400: string;
        500: string;
        600: string;
    };
    yellow: {
        300: string;
        600: string;
    };
};